move firstPublished and lastUpdated to common
